--- scrs/src/main.rs
@@ -1,5 +1,0 @@
-//use search_compiler_native::hello;
-
-fn main() {
-    println!("Hi"); 
-}
